Estiliza tus ideas first version.
Description:
Is a landing page, simple and estethic. Because it doesnt need to collect data (no get methods  with data from the client)
it was construct in nodejs with Nextjs, because it will allowed me to render the page once in the server and
improve the perfomance.

Design by Victoria Iztayana.

Things i need to improve are:
	i) details in the responsive of the web
	ii) the overscroll hidden in the body when i press the toggle hamb in mobile version
	iii) maybe the animation of "Estiliza tus ideas"
	iv) change to static site generator for better performance